Update neutron.json (incorrect ibc/denom) (#357)
* Update neutron.json ( incorrect ibc/denom

So the previous ibc/denom was the token being ibcd from chihuahua chain to osmosis and then to astroport. This ibc/denom is just from chihuahua chain to astroport. Thanks for the review.

* Update neutron.json

changed current baddog, added original baddog and made it clear which was the (osmosis) baddog. Thank you for the review.

* Update neutron.json

change placement